+++ packages/modules/itwin-viewer-react/src/components/iModel/IModelLoader.tsx
@@ -6,12 +6,16 @@
 import "@bentley/icons-generic-webfont/dist/bentley-icons-generic-webfont.css";
 import "./IModelLoader.scss";
 
+import { ClientRequestContext, Config } from "@bentley/bentleyjs-core";
 import {
   IModelApp,
   IModelConnection,
+  MessageBoxIconType,
+  MessageBoxType,
   SnapshotConnection,
   ViewState,
 } from "@bentley/imodeljs-frontend";
+import { UrlDiscoveryClient } from "@bentley/itwin-client";
 import { useErrorManager } from "@bentley/itwin-error-handling-react";
 import {
   BackstageActionItem,
@@ -35,6 +39,7 @@ import {
 } from "../../services/iModel/IModelService";
 import { SelectionScopeClient } from "../../services/iModel/SelectionScopeClient";
 import { ViewCreator } from "../../services/iModel/ViewCreator";
+import Initializer from "../../services/Initializer";
 import { ai } from "../../services/telemetry/TelemetryService";
 import {
   ItwinViewerUi,
@@ -126,14 +131,37 @@ const Loader: React.FC<ModelLoaderProps> = React.memo(
           imodelConnection = await openImodel(contextId, iModelId, changeSetId);
         }
         if (imodelConnection) {
+          // Tell the SyncUiEventDispatcher and StateManager about the iModelConnection
+          UiFramework.setIModelConnection(imodelConnection);
+          SyncUiEventDispatcher.initializeConnectionEvents(imodelConnection);
+
           if (onIModelConnected) {
             onIModelConnected(imodelConnection);
           }
-          // TODO revist this logic for the viewer
-          // pass the default viewids to the frontstage.
-          // currently we pass the first 2 spatial views to support split screen
-          // this logic will likely change when we have proper use cases
+
           const viewIds = await getDefaultViewIds(imodelConnection);
+
+          if (viewIds.length === 0 && contextId && iModelId) {
+            // no valid view data in the model. Direct the user to the synchronization portal
+            const msgDiv = document.createElement("div");
+            const msg = await Initializer.getIModelDataErrorMessage(
+              contextId,
+              iModelId,
+              IModelApp.i18n.translateWithNamespace(
+                "iTwinViewer",
+                "iModels.emptyIModelError"
+              )
+            );
+            msgDiv.innerHTML = msg;
+            // this can and should be async. No need to wait on it
+            // eslint-disable-next-line @typescript-eslint/no-floating-promises
+            IModelApp.notifications.openMessageBox(
+              MessageBoxType.Ok,
+              msgDiv,
+              MessageBoxIconType.Critical
+            );
+          }
+
           // attempt to construct a default viewState
           const savedViewState = await ViewCreator.createDefaultView(
             imodelConnection,
@@ -146,12 +174,8 @@ const Loader: React.FC<ModelLoaderProps> = React.memo(
             throw new Error("No default view state for the imodel!");
           }
 
-          // Tell the SyncUiEventDispatcher about the iModelConnection
-          UiFramework.setIModelConnection(imodelConnection);
           // Set default view state
           UiFramework.setDefaultViewState(savedViewState);
-
-          SyncUiEventDispatcher.initializeConnectionEvents(imodelConnection);
 
           // TODO revist for snapshots once settings are removed
           if (!snapshotPath) {
@@ -246,7 +270,10 @@ const Loader: React.FC<ModelLoaderProps> = React.memo(
     if (error) {
       throw error;
     } else {
-      return finalFrontstages && finalBackstageItems && connected ? (
+      return finalFrontstages &&
+        finalBackstageItems &&
+        connected &&
+        StateManager.store ? (
         <div className="itwin-viewer-container">
           <Provider store={StateManager.store}>
             <IModelViewer